refactor: update project structure and organization
