--- src/hooks/usePasswordStrength/index.ts
@@ -1,2 +1,0 @@
-export { usePasswordStrength } from "./usePasswordStrength";
-export type * from "./types";
