+++ src/index.ts
@@ -1,4 +1,3 @@
-export { usePasswordStrength } from "./hooks/usePasswordStrength/index";
-export type * from "./hooks/usePasswordStrength/index";
-export { validatePasswordStrength } from "./validatePasswordStrength";
-export type * from "./validatePasswordStrength";
+export * from "./validatePasswordStrength";
+export * from "./usePasswordStrength";
+export type * from "./types";
