+++ src/types/passwordValidation.ts
@@ -1,3 +1,21 @@
+export interface IValidatePasswordStrengthResponse {
+  validationResult: IValidationRule[];
+  score: number;
+  entropy: number;
+  isValid: boolean;
+  password: string;
+}
+
+export type IValidatePasswordStrengthModes = "strict" | "regex" | "score";
+
+export interface IValidatePasswordStrengthOptions {
+  maxScore?: number;
+  configMessages?: IValidationMessages;
+  mode?: IValidatePasswordStrengthModes;
+  minBestEntropy?: number;
+  minRequiredScore?: number;
+}
+
 export interface IUsePasswordStrengthActions {
   validationResult: IValidationRule[];
   score: number;
