+++ src/validatePasswordStrength.ts
@@ -1,5 +1,5 @@
-import { calculateEntropy, calculatePasswordScore, validatePasswordRules } from "../utils/passwordStrengthUtils";
 import { IValidatePasswordStrengthOptions, IValidatePasswordStrengthResponse } from "./types";
+import { calculateEntropy, calculatePasswordScore, validatePasswordRules } from "./utils";
 
 const defaultMessages = {
   minLowercaseMessage: "At least 1 lowercase letter",
