--- src/validatePasswordStrength/index.ts
@@ -1,2 +1,0 @@
-export { validatePasswordStrength } from "./validatePasswordStrength";
-export type * from "./types";
